Remove auth action. Extract window class
